% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalizeObject.R
\name{finalizeObject}
\alias{finalizeObject}
\title{Finalize an object.}
\usage{
finalizeObject(object, publicMethodNames)
}
\arguments{
\item{object}{an object}

\item{publicMethodNames}{a character vector of all public method names}
}
\description{
This function should be called at the end of a constructor function to finalize an object.
}
\details{
See http://www.github.com/cartowong/oo.
}
