% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{include}
\alias{include}
\title{Source all R script files under the given path, recursively.}
\usage{
include(path)
}
\arguments{
\item{path}{a string. The path of a directory in which all R script files will be sourced.}
}
\value{
a vector of strings. The paths of the sourced files.
}
\description{
Source all R script files under the given path, recursively.
}
