% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precondition.R
\name{Precondition}
\alias{Precondition}
\title{Constructor of Precondition}
\usage{
Precondition(allowNA = FALSE, allowNull = FALSE)
}
\arguments{
\item{allowNA}{Should NA be allowed? The method checkIsNull ignores this parameter.}

\item{allowNull}{Should NULL be allowed? The method checkIsNA ignores this parameter.}
}
\value{
an instance of Precondition
}
\description{
Utility functions for checking the type of an argument, based on TypeChecker in the same package.
If the argument is not of the expected type, an error will be thrown. Usually, these utility
functions are called at the begining of a function body.
}
\examples{
precondition <- Precondition()
myFunction <- function(x) {
  precondition$isNumeric(x, 'x should be numeric')
  ...
}

}
