% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitTester.R
\name{UnitTester}
\alias{UnitTester}
\title{Constructor of the class UnitTester.}
\usage{
UnitTester(turnWarningsToErrors = TRUE, hidePassed = FALSE)
}
\value{
an instance of UnitTester
}
\description{
Constructor of the class UnitTester.
}
\examples{
unitTester <- UnitTester()

unitTester$addTest('Test 1', function() {
  unitTester$assertEqual(1, 1)
})

unitTester$addTest('Test 2', function() {
  unitTester$assertThrow(function() {
    stop('Error from Test 2')
  })
})

unitTester$runAllTests()

}
