% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precondition.R
\docType{data}
\name{Precondition}
\alias{Precondition}
\title{Precondition}
\format{An object of class \code{list} of length 6.}
\usage{
Precondition
}
\description{
Utility functions for checking the type of an argument, based on TypeChecker in the same package.
If the argument is not of the expected type, an error will be thrown. Usually, these utility
functions are called at the begining of a function body.
}
\examples{
myFunction <- function(x) {
  Precondition$isNumeric(x, 'x is not numeric')
  ...
}

}
\keyword{datasets}
