% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypeChecker.R
\name{TypeChecker}
\alias{TypeChecker}
\title{Constructor of the class TypeChecker}
\usage{
TypeChecker(allowNA = FALSE, allowNull = FALSE)
}
\arguments{
\item{allowNA}{Should NA be allowed? The method isNull ignores this parameter.}

\item{allowNull}{Should NULL be allowed? The method isNA ignores this parameter.}
}
\value{
an instance of TypeChecker
}
\description{
Utility functions for checking the type of an object. Unlike, for example, is.numeric in the base
package, the function isNumeric in TypeChecker checks whether the given object is a single numeric
value. A vector of two numeric values is not considered as numeric type. Conceptually, its type
is vector whose element type happens to be numeric.
}
\examples{
TypeChecker()$isNA(x)
TypeChecker()$isNull(x)
TypeChecker()$isBoolean(x)
TypeChecker()$isFunction(x)
TypeChecker()$isNumeric(x)
TypeChecker()$isString(x)

}
