% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitTester.R
\name{UnitTester}
\alias{UnitTester}
\title{Constructor of the class UnitTester.}
\usage{
UnitTester(turnWarningsToErrors = TRUE)
}
\value{
an instance of UnitTester
}
\description{
Constructor of the class UnitTester.
}
