% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitTester.R
\name{UnitTester}
\alias{UnitTester}
\title{Constructor of the class UnitTester}
\usage{
UnitTester(turnWarningsToErrors = TRUE, hidePassed = FALSE)
}
\arguments{
\item{turnWarningsToErrors}{If set to TRUE, all warnings issued within any test method are turned into errors. The default value is TRUE.}

\item{hidePassed}{If set to TRUE, all passed test results are hidden in the output. The is useful when you are only interested in the failed assertions and the unexpected errors. The default value is FALSE.}
}
\value{
an instance of UnitTester
}
\description{
Constructor of the class UnitTester
}
\examples{
unitTester <- UnitTester()

unitTester$addTest('Test 1', function() {
  unitTester$assertEqual(1, 1)
})

unitTester$addTest('Test 2', function() {
  unitTester$assertThrow(function() {
    stop('Error from Test 2')
  })
})

unitTester$runAllTests()

}
